♻️ refactor metrics page
